refactor(modules): move auth controller inside account module
