+++ src/account/account.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
-import { AccountController } from './account.controller';
+import { AccountController } from './controllers/account.controller';
 import { AuthModule } from '../auth/auth.module';
+import { AuthController } from './controllers/auth.controller';
 
 @Module({
     imports: [AuthModule],
-    controllers: [AccountController],
+    controllers: [AccountController, AuthController],
 })
 export class AccountModule {}
