+++ src/account/tests/e2e/account.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AccountController } from './account.controller';
+import { AccountController } from '../../controllers/account.controller';
 
 describe('AccountController', () => {
     let controller: AccountController;
